move static/dynamic abi types
