+++ packages/common/src/utils/index.ts
@@ -1,1 +1,2 @@
+export * from "./assertExhaustive";
 export * from "./curry";
