+++ packages/protocol-parser/src/abiTypesToSchema.ts
@@ -1,6 +1,5 @@
-import { StaticAbiType, DynamicAbiType } from "@latticexyz/schema-type";
+import { StaticAbiType, DynamicAbiType, staticAbiTypeToByteLength } from "@latticexyz/schema-type";
 import { Schema } from "./common";
-import { staticAbiTypeToByteLength } from "./staticAbiTypes";
 import { abiTypesToSchemaData } from "./abiTypesToSchemaData";
 
 export function abiTypesToSchema(staticFields: StaticAbiType[], dynamicFields: DynamicAbiType[] = []): Schema {
