+++ packages/protocol-parser/src/abiTypesToSchemaData.ts
@@ -1,6 +1,5 @@
 import { Hex } from "viem";
-import { schemaAbiTypes, StaticAbiType, DynamicAbiType } from "@latticexyz/schema-type";
-import { staticAbiTypeToByteLength } from "./staticAbiTypes";
+import { schemaAbiTypes, StaticAbiType, DynamicAbiType, staticAbiTypeToByteLength } from "@latticexyz/schema-type";
 
 export function abiTypesToSchemaData(staticFields: StaticAbiType[], dynamicFields: DynamicAbiType[] = []): Hex {
   const staticDataLength = staticFields.reduce((acc, fieldType) => acc + staticAbiTypeToByteLength[fieldType], 0);
