+++ packages/protocol-parser/src/decodeKeyTuple.ts
@@ -1,6 +1,6 @@
-import { Schema } from "./common";
 import { Hex, decodeAbiParameters } from "viem";
-import { StaticPrimitiveType } from "./staticAbiTypes";
+import { StaticPrimitiveType } from "@latticexyz/schema-type";
+import { Schema } from "./common";
 
 // key tuples are encoded in the same way as abi.encode, so we can decode them with viem
