+++ packages/protocol-parser/src/index.ts
@@ -1,13 +1,10 @@
 export * from "./abiTypesToSchema";
 export * from "./abiTypesToSchemaData";
-export * from "./abiTypeToDefaultValue";
 export * from "./common";
 export * from "./decodeDynamicField";
 export * from "./decodeKeyTuple";
 export * from "./decodeStaticField";
-export * from "./dynamicAbiTypes";
 export * from "./errors";
 export * from "./hexToPackedCounter";
 export * from "./hexToSchema";
 export * from "./hexToTableSchema";
-export * from "./staticAbiTypes";
