+++ packages/schema-type/package.json
@@ -32,7 +32,8 @@
     "test": "vitest typecheck --run && vitest --run && forge test"
   },
   "dependencies": {
-    "abitype": "0.8.7"
+    "abitype": "0.8.7",
+    "viem": "1.0.6"
   },
   "devDependencies": {
     "ds-test": "https://github.com/dapphub/ds-test.git#c9ce3f25bde29fc5eb9901842bf02850dfd2d084",
