+++ packages/schema-type/src/typescript/index.ts
@@ -1,1 +1,4 @@
+export * from "./dynamicAbiTypes";
 export * from "./schemaAbiTypes";
+export * from "./schemaAbiTypeToDefaultValue";
+export * from "./staticAbiTypes";
