+++ packages/schema-type/src/typescript/schemaAbiTypeToDefaultValue.ts
@@ -1,8 +1,8 @@
-import { SchemaAbiType } from "@latticexyz/schema-type";
+import { SchemaAbiType } from "./schemaAbiTypes";
 import { DynamicAbiTypeToPrimitiveType, dynamicAbiTypeToDefaultValue } from "./dynamicAbiTypes";
 import { StaticAbiTypeToPrimitiveType, staticAbiTypeToDefaultValue } from "./staticAbiTypes";
 
-export const abiTypeToDefaultValue = {
+export const schemaAbiTypeToDefaultValue = {
   ...staticAbiTypeToDefaultValue,
   ...dynamicAbiTypeToDefaultValue,
 } as const satisfies Record<SchemaAbiType, StaticAbiTypeToPrimitiveType | DynamicAbiTypeToPrimitiveType>;
